+++ taskflowbackend/flows/base_flow.py
@@ -1,9 +1,14 @@
 import logging
 
+from typing import Any, Optional
+
 from taskflow import engines
 from taskflow.patterns import linear_flow as lf
 
 from django.conf import settings
+
+# Projectroles dependency
+from projectroles.models import Project, SODARUser
 
 from taskflowbackend.tasks.base_task import BaseTask, ForceFailException
 
@@ -16,17 +21,19 @@ class BaseLinearFlow:
 
     def __init__(
         self,
-        irods_backend,
-        project,
-        flow_name,
-        flow_data,
-        async_mode=False,
-        tl_event=None,
+        irods_backend: Any,
+        project: Project,
+        user: Optional[SODARUser],
+        flow_name: str,
+        flow_data: Optional[dict],
+        async_mode: bool = False,
+        tl_event: Any = None,
     ):
         self.irods_backend = irods_backend
         self.irods = irods_backend.get_session_obj()
         self.irods.connection_timeout = settings.TASKFLOW_IRODS_CONN_TIMEOUT
         self.project = project
+        self.user = user
         self.flow_name = flow_name
         self.flow_data = flow_data
         self.required_fields = []  # For validation